backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
pydantic-settings
sse-starlette
